Implemented "New" Notifications and Toggle Notifications

+++ ggc-core/src/ggc/BargainNotification.java
@@ -2,7 +2,7 @@ package ggc;
 
 public class BargainNotification extends Notification{
 
-    public BargainNotification(String productId, int productPrice){
+    public BargainNotification(String productId, double productPrice){
         super(productId, productPrice);
     }
